feat: 조회 방식 수정

+++ backend/src/main/java/com/quiet/onterview/community/dto/response/CommentListResponse.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 @Builder
 public class CommentListResponse {
 
-    Long commentCount;
+    Integer commentCount;
     List<CommentObjectResponse> comments;
 }
